+++ tests/WordFormatter/LeetDecoderTest.php
@@ -9,21 +9,21 @@ use Stadly\PasswordPolice\WordFormatter;
 use Traversable;
 
 /**
- * @coversDefaultClass \Stadly\PasswordPolice\WordFormatter\Leetspeak
+ * @coversDefaultClass \Stadly\PasswordPolice\WordFormatter\LeetDecoder
  * @covers ::<protected>
  * @covers ::<private>
  */
-final class LeetspeakTest extends TestCase
+final class LeetDecoderTest extends TestCase
 {
     /**
      * @covers ::__construct
      */
     public function testCanConstructFormatter(): void
     {
-        $formatter = new Leetspeak();
+        $formatter = new LeetDecoder();
 
         // Force generation of code coverage
-        $formatterConstruct = new Leetspeak();
+        $formatterConstruct = new LeetDecoder();
         self::assertEquals($formatter, $formatterConstruct);
     }
 
@@ -32,7 +32,7 @@ final class LeetspeakTest extends TestCase
      */
     public function testCanFormatWordWithoutLeetspeak(): void
     {
-        $formatter = new Leetspeak();
+        $formatter = new LeetDecoder();
 
         self::assertSame(['fOoBaR'], iterator_to_array($formatter->apply(['fOoBaR']), false));
     }
@@ -42,7 +42,7 @@ final class LeetspeakTest extends TestCase
      */
     public function testCanFormatWordsWithoutLeetspeak(): void
     {
-        $formatter = new Leetspeak();
+        $formatter = new LeetDecoder();
 
         self::assertSame(['fOo', 'BaR'], iterator_to_array($formatter->apply(['fOo', 'BaR']), false));
     }
@@ -52,7 +52,7 @@ final class LeetspeakTest extends TestCase
      */
     public function testCanFormatWordWithLeetspeak(): void
     {
-        $formatter = new Leetspeak();
+        $formatter = new LeetDecoder();
 
         self::assertContains('LEET SPEAK', iterator_to_array($formatter->apply(['1337 5P34K']), false));
     }
@@ -62,7 +62,7 @@ final class LeetspeakTest extends TestCase
      */
     public function testCanFormatWordsWithLeetspeak(): void
     {
-        $formatter = new Leetspeak();
+        $formatter = new LeetDecoder();
 
         self::assertContains('LEET', iterator_to_array($formatter->apply(['1337', '5P34K']), false));
         self::assertContains('SPEAK', iterator_to_array($formatter->apply(['1337', '5P34K']), false));
@@ -73,7 +73,7 @@ final class LeetspeakTest extends TestCase
      */
     public function testCanApplyFormatterChain(): void
     {
-        $formatter = new Leetspeak();
+        $formatter = new LeetDecoder();
 
         $next = $this->createMock(WordFormatter::class);
         $next->method('apply')->willReturnCallback(